+++ azzeroCO2-war/src/main/java/it/agilis/mens/azzeroCO2/shared/model/evento/TrasportoPersoneModel.java
@@ -11,8 +11,8 @@ import com.extjs.gxt.ui.client.data.BaseModel;
  */
 public class TrasportoPersoneModel extends BaseModel {
 
-     public long getId()  {
-        return (Long) get("id");
+     public Long getId()  {
+        return get("id");
     }
     public void setId(Long id){
         set("id",id);
